requirements.txt
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
Werkzeug==0.14.1
Flask-WTF==0.14.2
WTForms==2.2.1
PyJWT
SQLAlchemy
Flask-SQLAlchemy